[Silver II] Title: 랜선 자르기, Time: 156 ms, Memory: 17212 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.java
@@ -27,7 +27,7 @@ public class Main {
         while(left <= right) {
             long mid = (left + right) / 2;
 			
-			if(canCut((int)mid)>= M) {
+			if(canCut(mid)>= M) {
 				ans = mid;
 				left = mid+1;
 			}else {
@@ -37,7 +37,7 @@ public class Main {
 		System.out.println(ans);
 		
 	}
-    static long canCut(int target) {
+    static long canCut(long target) {
         long sum = 0;
 		for(int i = 0; i<N; i++) {
             sum += arr[i]/ target;
